+++ src/Alpacka.Lib/AlpackaInfo.cs
@@ -1,5 +1,7 @@
 using System.IO;
 using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
+using Newtonsoft.Json.Serialization;
 
 namespace Alpacka.Lib
 {
