--- src/GitMC.Lib/Git/UpdateUtil.cs
@@ -1,18 +1,0 @@
-using System;
-using System.IO;
-using LibGit2Sharp;
-
-namespace GitMC.Lib.Git
-{
-    public static class UpdateUtil
-    {
-        public static string Checkout(string workingDirectory, string branch)
-        {
-            // var directory = Path.Combine(workingDirectory, $"{ Guid.NewGuid() }");
-            // Repository.Clone(url, directory);
-            
-            // return branch;
-            return "";
-        }
-    }
-}
